moved language detection core to its own file
