+++ app/core.js
@@ -883,50 +883,9 @@ exports.runNode = function (req, res, io) {
 					switch (node.subtype) {
 						
 						case "language":
-							// find everything
-							mongoquery.find2({}, node.collection, function(err, docs) {
-								sandbox.context.doc_count = docs.length;
-								console.log(docs.length);
-								runNodeScriptInContext("init", node, sandbox, io);
-								var cld = require("cld");
-								
-								// run node once per record
-								async.eachSeries(docs, function iterator(doc, next) {
-
-									sandbox.context.doc = doc;
-									sandbox.context.count++;
-									sandbox.out.value = null;  // reset output
-									var input = node.params.in_field;
-									
-									if (doc[input] && doc[input].constructor.name == "Array") {
-										async.eachSeries(doc[input], function iterator(value, nextvalue) {
-											sandbox.context.data = [];
-											cld.detect(value, function(err, result) {
-												sandbox.context.data.push(result);
-												console.log("LANG:",result);
-												nextvalue();
-											});
-										}, function done() {
-											run.runInContext(sand);
-											var setter = {};
-											setter[node.out_field] = sandbox.out.value;
-											console.log("SETTER:", setter);
-											mongoquery.update(node.collection, {_id:sandbox.context.doc._id},{$set:setter}, next);
-										})
-										
-									} else {
-										console.log("ABSTRACT NOT FOUND");
-										next();
-									}
-
-									
-
-									
-								}, function done () {
-									runNodeScriptInContext("finish", node, sandbox, io);
-									//exports.updateView(node, sandbox, io, function(msg) {console.log("NODE: view created", msg);});
-								});
-							});
+						
+							var detect = require("../app/node_runners/field-detect-language.js");
+							asyncLoop.loop(node, sandbox, detect.language);
 							break;
 						}
